# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  customerMaster.CustomerMaster:
    properties:
      create_date:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  customerMaster.CustomerMasterViewModel:
    properties:
      name:
        type: string
    type: object
  customerMember.CustomerMember:
    properties:
      create_date:
        type: string
      customer_group_entity_relation_id:
        type: integer
      id:
        type: integer
      user_info_userid:
        type: string
    type: object
  customerMember.CustomerMemberViewModel:
    properties:
      customer_group_entity_relation_id:
        type: integer
      user_info_userid:
        type: string
    type: object
  user.CreateUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      login_user_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  user.User:
    properties:
      active_status:
        type: boolean
      email:
        type: string
      first_name:
        type: string
      is_email_confirmed:
        type: boolean
      last_name:
        type: string
      login_user_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      registration_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact:
    email: chandresh@techpanion.com
    name: API Support
    url: http://www.techpanion.com/support
  description: This is a sample server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Techpanion Admin Portal API
  version: "1.0"
paths:
...
  /customerMaster/:
    get:
      consumes:
      - application/json
      description: Here Get customer by id
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customerMaster.CustomerMaster'
            type: array
        "400": {}
        "404": {}
        "500": {}
      summary: Get Customer
      tags:
      - CustomerMaster
    post:
      consumes:
      - application/json
      description: Here Save customer info.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Customer Details
        in: body
        name: cMaster
        required: true
        schema:
          $ref: '#/definitions/customerMaster.CustomerMasterViewModel'
      produces:
      - application/json
      responses:
        "200": {}
        "400": {}
        "404": {}
        "409": {}
        "500": {}
      summary: Save Customer Info
      tags:
      - CustomerMaster
  /customerMaster/{id}:
    delete:
      consumes:
      - application/json
      description: Here we can remve customer
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Customer Master ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
        "400": {}
        "404": {}
        "500": {}
      summary: delete particular customer
      tags:
      - CustomerMaster
  /customerMember/:
    post:
      consumes:
      - application/json
      description: Here Save customer info.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Customer Details
        in: body
        name: custM
        required: true
        schema:
          $ref: '#/definitions/customerMember.CustomerMemberViewModel'
      produces:
      - application/json
      responses:
        "200": {}
        "400": {}
        "404": {}
        "500": {}
      summary: Save Customer Info
      tags:
      - CustomerMember
  /customerMember/{id}:
    delete:
      consumes:
      - application/json
      description: Here we can remove particular customer
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Customer Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
        "400": {}
        "404": {}
        "500": {}
      summary: delete particular customer
      tags:
      - CustomerMember
    get:
      consumes:
      - application/json
      description: Here Get customer by id
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customerMember.CustomerMember'
            type: array
        "400": {}
        "404": {}
        "500": {}
      summary: Get Customer
      tags:
      - CustomerMember
  /users/:
    post:
      consumes:
      - application/json
      description: Please create the user for providing permission
      parameters:
      - description: Create User Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUser'
      produces:
      - application/json
      responses:
        "201": {}
        "400": {}
        "404": {}
        "500": {}
      summary: Create User
      tags:
      - User
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Here we will get particular user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400": {}
        "404": {}
        "500": {}
      summary: Get Particular User
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: /access_token/
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: /access_token/
    type: oauth2
  OAuth2Application:
    flow: application
    tokenUrl: /access_token/
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: /access_token/
    flow: implicit
    type: oauth2
  OAuth2Password:
    flow: password
    tokenUrl: /access_token/
    type: oauth2
swagger: "2.0"