# pkg/customerMember/customerMemberViewModel.go
package customerMember

type CustomerMemberViewModel struct {
	CustomerGroupEntityRelationID int64  `json:"customer_group_entity_relation_id"`
	UserInfoUserID                string `json:"user_info_userid"`
}